backend/requirements.txt
# Core Framework
Flask==3.0.0
Flask-CORS==4.0.0

# Data Validation
pydantic==2.5.0

# HTTP Client
requests==2.31.0

# Caching
cachetools==5.3.2

# Environment Variables
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code Quality
flake8==6.1.0
black==23.12.1
pylint==3.0.3

# Type Checking
mypy==1.7.1

# Date/Time
python-dateutil==2.8.2
